# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hackathon Strategy Framework API
  version: 1.0.0
paths:
...
  /api/v1/documents/:
    get:
      tags:
      - Document Vault & Search
      summary: List Documents
      description: Lists indexed document chunks from PostgreSQL or demo storage.
      operationId: list_documents_api_v1_documents__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_list_DocumentChunkResponse__'
    post:
      tags:
      - Document Vault & Search
      summary: Create Document
      description: Ingests and indexes a new document chunk with vector embedding
        support.
      operationId: create_document_api_v1_documents__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentChunkCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_DocumentChunkResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/search:
    post:
      tags:
      - Document Vault & Search
      summary: Search Documents
      description: 'Executes hybrid dense + sparse vector search with Reciprocal Rank
        Fusion (RRF)

        and optional FlashRank cross-encoder reranking.'
      operationId: search_documents_api_v1_documents_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_list_DocumentSearchResult__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DocumentChunkCreate:
      properties:
        title:
          type: string
          maxLength: 255
          title: Title
          description: Document or section title
        content:
          type: string
          title: Content
          description: Raw text content of the chunk
        source:
          type: string
          maxLength: 255
          title: Source
          description: Source provenance (e.g., pdf, web, manual)
          default: manual
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
          description: Optional pre-computed 1536-dim embedding vector
      type: object
      required:
      - title
      - content
      title: DocumentChunkCreate
    DocumentChunkResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          type: string
          title: Content
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          default: manual
        created_at:
          anyOf:
          - {}
          - type: 'null'
          title: Created At
      type: object
      required:
      - id
      - content
      title: DocumentChunkResponse
    DocumentSearchRequest:
      properties:
        query:
          type: string
          minLength: 1
          title: Query
          description: Search query string
        top_k:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Top K
          description: Number of results to return
          default: 5
        use_rerank:
          type: boolean
          title: Use Rerank
          description: Whether to apply neural cross-encoder reranking
          default: true
      type: object
      required:
      - query
      title: DocumentSearchRequest
    DocumentSearchResult:
      properties:
        id:
          type: integer
          title: Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          type: string
          title: Content
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        score:
          type: number
          title: Score
        rank:
          type: integer
          title: Rank
      type: object
      required:
      - id
      - content
      - score
      - rank
      title: DocumentSearchResult
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ResponseEnvelope_DocumentChunkResponse_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          anyOf:
          - $ref: '#/components/schemas/DocumentChunkResponse'
          - type: 'null'
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          default: Operation successful
      type: object
      title: ResponseEnvelope[DocumentChunkResponse]
...
    ResponseEnvelope_list_DocumentChunkResponse__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/DocumentChunkResponse'
            type: array
          - type: 'null'
          title: Data
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          default: Operation successful
      type: object
      title: ResponseEnvelope[list[DocumentChunkResponse]]
    ResponseEnvelope_list_DocumentSearchResult__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/DocumentSearchResult'
            type: array
          - type: 'null'
          title: Data
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          default: Operation successful
      type: object
      title: ResponseEnvelope[list[DocumentSearchResult]]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError